Move Photoelectrons outside of the event module
