+++ sstcam_simulation/__init__.py
@@ -1,5 +1,6 @@
+from .photoelectrons import Photoelectrons
 from .camera import Camera, SSTCameraMapping
-from .event import PhotoelectronSource, EventAcquisition, Photoelectrons
+from .event import PhotoelectronSource, EventAcquisition
 from .io import SimtelReader, PhotoelectronWriter, PhotoelectronReader
 
 __version__ = "2.1.0"
