+++ sstcam_simulation/camera/tests/test_pulse.py
@@ -1,6 +1,7 @@
 from sstcam_simulation.camera.pulse import PhotoelectronPulse, GenericPulse, \
     ReferencePulse
 from sstcam_simulation.camera.constants import CONTINUOUS_READOUT_SAMPLE_WIDTH
+from inspect import isabstract
 import numpy as np
 from scipy.stats import norm
 import pytest
@@ -10,12 +11,12 @@ classes = PhotoelectronPulse.__subclasses__()
 # GenericPulse must be tested separately
 classes.remove(GenericPulse)
 
-# TODO: remove post-deprecation completion
-classes.remove(ReferencePulse)
-
 
 @pytest.mark.parametrize("ref_pulse_class", classes)
 def test_reference_pulses(ref_pulse_class):
+    if isabstract(ref_pulse_class):
+        return
+
     pulse = ref_pulse_class()
     np.testing.assert_allclose(pulse.area, 1)
     assert pulse.time.size == pulse.amplitude.size
@@ -23,6 +24,9 @@ def test_reference_pulses(ref_pulse_class):
 
 @pytest.mark.parametrize("ref_pulse_class", classes)
 def test_reference_pulses_mv_per_pe(ref_pulse_class):
+    if isabstract(ref_pulse_class):
+        return
+
     pulse = ref_pulse_class(mv_per_pe=2)
     np.testing.assert_allclose(pulse.height, 2)
     assert pulse.time.size == pulse.amplitude.size
