+++ sstcam_simulation/event/__init__.py
@@ -1,3 +1,2 @@
-from .photoelectrons import Photoelectrons
 from .source import PhotoelectronSource
 from .acquisition import EventAcquisition
