+++ sstcam_simulation/io/tests/test_photoelectron_writer.py
@@ -1,5 +1,5 @@
 from sstcam_simulation.io.photoelectron_writer import PhotoelectronWriter
-from sstcam_simulation.event.photoelectrons import Photoelectrons
+from sstcam_simulation.photoelectrons import Photoelectrons
 import tables
 import numpy as np
 import pytest
